Extract game-phase Divider into reviewdivision.ts
